+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## not released
 
+## v1.3.0 (2023-05-14)
+
+- Add: German translation
+
+## v1.2.2 (2023-01-06)
+
+- Fix: `moveLogFile: ENOTDIR: not a directory` for certain backup settings
+- Add: DirectoryPath selector for backup path and tmp. export path selection on Joplin >= v2.10.4
+
 ## v1.2.1 (2022-12-10)
 
 - Fix: #47 Suppress repeating error message during automatic execution if the backup path is not accessible